cycles 134+135: extract BandMaskTab + ApplyToDocumentTab + RecipesTab + QKExploreTab + StatCard into modules
Continues the Workspace.tsx modularisation started in c133. Four
self-contained tab components (and their sub-components) moved out
of the monolith into pages/workspace/tabs/<name>.tsx:

- BandMaskTab.tsx (565 LOC) — c126/127 Step 8 band-mask sweep:
  4-card picker + BandMaskComparisonOverview table + BandMaskDetailCard
  with top-words + P(L|t) + canonical-comparison panel
- ApplyToDocumentTab.tsx (405 LOC) — c101 Step 7: DocPicker filter
  + sortable table + DocDetailPanel ranked θ vs marginal +
  PerTopicLabelBars; useMemoMarginal hook
- RecipesTab.tsx (191 LOC) — c102 V1..V12 × scheme × Q corpus
  inspector: RecipeAxisPicker + RecipeDetailCard with doc-length
  distribution + top tokens
- QKExploreTab.tsx (396 LOC) — c107 K-sweep observatory:
  recommended_K chip + 3 SVG curves (perplexity / topic_diversity /
  matched_cosine) + per-K detail table; SweepCurveCard

Also extracted the shared StatCard (was inline UnmixingStat in
Workspace.tsx, used by ~6 tabs) into pages/workspace/components/
with the original name preserved as an export alias.

Workspace.tsx shrank from 11461 → 9896 LOC (-1565 across c133-c135).
Build clean, no behaviour change.

Refs #421
