+++ frontend/src/pages/Workspace.tsx
@@ -36,6 +36,11 @@ import {
   type RoutedPrediction,
   computeRoutedPrediction,
 } from "./workspace/helpers/routedPrediction";
+import { BandMaskTab } from "./workspace/tabs/BandMaskTab";
+import { ApplyToDocumentTab } from "./workspace/tabs/ApplyToDocumentTab";
+import { RecipesTab } from "./workspace/tabs/RecipesTab";
+import { QKExploreTab } from "./workspace/tabs/QKExploreTab";
+import { UnmixingStat } from "./workspace/components/StatCard";
 
 const Scatter3D = lazy(() =>
   import("@/components/plots/Scatter3D").then((m) => ({ default: m.Scatter3D })),
@@ -7226,18 +7231,6 @@ function UnmixingTab({
   );
 }
 
-function UnmixingStat({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="flex flex-col">
-      <span className="text-[10px] uppercase tracking-widest font-semibold" style={{ color: "var(--color-fg-faint)" }}>
-        {label}
-      </span>
-      <span className="text-[22px] font-mono leading-tight" style={{ color: "var(--color-fg)" }}>
-        {value}
-      </span>
-    </div>
-  );
-}
 
 function UnmixingSpectraCard({
   title,
@@ -9900,1561 +9893,4 @@ function KeyboardShortcutsOverlay({ onClose }: { onClose: () => void }) {
   );
 }
 
-function ApplyToDocumentTab({
-  isLoading,
-  error,
-  docs,
-  topicViews,
-  topicToData,
-  selectedTopic,
-  setSelectedTopic,
-}: {
-  isLoading: boolean;
-  error: Error | null;
-  docs: import("@/api/client").DocumentCardsFile | null;
-  topicViews: import("@/api/client").TopicViews | null;
-  topicToData: import("@/api/client").TopicToData | null;
-  selectedTopic: number | null;
-  setSelectedTopic: (k: number | null) => void;
-}) {
-  const [pickedIdx, setPickedIdx] = useState<number | null>(null);
-  const docCards = docs?.document_cards ?? [];
-  const K = topicViews?.topic_count ?? 12;
-  const marginalTheta = useMemoMarginal(docCards, K);
 
-  if (isLoading) return <p style={{ color: "var(--color-fg-faint)" }}>Loading document cards…</p>;
-  if (error) {
-    return (
-      <div className="rounded-lg border p-6" style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)" }}>
-        <p style={{ color: "var(--color-warn)" }}>Could not load document data.</p>
-        <p className="mt-2 text-sm" style={{ color: "var(--color-fg-faint)" }}>{error.message}</p>
-      </div>
-    );
-  }
-  if (!docs || !topicViews) return null;
-
-  const picked = pickedIdx != null ? docCards[pickedIdx] : null;
-
-  return (
-    <div className="space-y-6">
-      <div
-        className="rounded-xl border p-5 relative overflow-hidden"
-        style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)", boxShadow: "var(--color-shadow)" }}
-      >
-        <div
-          aria-hidden
-          className="absolute top-0 left-0 right-0 h-1"
-          style={{ background: "linear-gradient(90deg, rgba(170,60,200,1) 0%, rgba(56,189,248,1) 100%)" }}
-        />
-        <h3 className="text-lg font-semibold tracking-tight mt-1 mb-1" style={{ color: "var(--color-fg)" }}>
-          Apply to a specific document (Step 7)
-        </h3>
-        <p className="text-[12.5px] mb-3" style={{ color: "var(--color-fg-faint)" }}>
-          Pick one of {docCards.length} sampled documents. The panels below show its full θ
-          vector (mixture over K = {K} topics), its dominant topic, its ground-truth label, and
-          how θ compares to the scene's marginal θ. Selecting a topic in any other tab carries
-          over here as a highlight.
-        </p>
-
-        <DocPicker
-          docCards={docCards}
-          pickedIdx={pickedIdx}
-          onPick={setPickedIdx}
-          K={K}
-        />
-      </div>
-
-      {picked ? (
-        <DocDetailPanel
-          doc={picked}
-          marginalTheta={marginalTheta}
-          topicViews={topicViews}
-          topicToData={topicToData}
-          selectedTopic={selectedTopic}
-          setSelectedTopic={setSelectedTopic}
-        />
-      ) : (
-        <div
-          className="rounded-lg border p-6 text-center text-[13px]"
-          style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)", color: "var(--color-fg-subtle)" }}
-        >
-          Pick a document from the table above to see its θ vector and label-vs-topic comparison.
-        </div>
-      )}
-    </div>
-  );
-}
-
-function useMemoMarginal(
-  docCards: import("@/api/client").DocumentCard[],
-  K: number,
-): number[] {
-  return useMemo(() => {
-    const sum = new Array(K).fill(0);
-    let count = 0;
-    for (const d of docCards) {
-      if (!d.theta_full) continue;
-      for (let k = 0; k < K; k++) sum[k] += d.theta_full[k] ?? 0;
-      count++;
-    }
-    if (count === 0) return sum;
-    return sum.map((v) => v / count);
-  }, [docCards, K]);
-}
-
-function DocPicker({
-  docCards,
-  pickedIdx,
-  onPick,
-  K,
-}: {
-  docCards: import("@/api/client").DocumentCard[];
-  pickedIdx: number | null;
-  onPick: (i: number) => void;
-  K: number;
-}) {
-  const [filterTopic, setFilterTopic] = useState<number | null>(null);
-  const [filterLabel, setFilterLabel] = useState<number | null>(null);
-
-  const filtered = useMemo(() => {
-    return docCards
-      .map((d, i) => ({ d, i }))
-      .filter(({ d }) => {
-        if (filterTopic !== null && d.topic_k_dominant !== filterTopic) return false;
-        if (filterLabel !== null && d.label_id !== filterLabel) return false;
-        return true;
-      });
-  }, [docCards, filterTopic, filterLabel]);
-
-  const topicSet = useMemo(() => {
-    return Array.from(new Set(docCards.map((d) => d.topic_k_dominant))).sort((a, b) => a - b);
-  }, [docCards]);
-  const labelSet = useMemo(() => {
-    const seen = new Map<number, string>();
-    for (const d of docCards) if (!seen.has(d.label_id)) seen.set(d.label_id, d.label_name);
-    return Array.from(seen.entries()).sort((a, b) => a[0] - b[0]);
-  }, [docCards]);
-
-  return (
-    <div className="space-y-3 mt-3">
-      <div className="flex items-baseline flex-wrap gap-3 text-[12px]" style={{ color: "var(--color-fg-faint)" }}>
-        <span className="uppercase tracking-widest font-semibold text-[10.5px]">Filter</span>
-        <span>by topic:</span>
-        <button
-          type="button"
-          onClick={() => setFilterTopic(null)}
-          className="rounded border px-1.5 py-0.5 text-[11px] font-mono"
-          style={{
-            borderColor: filterTopic === null ? "var(--color-accent)" : "var(--color-border)",
-            color: filterTopic === null ? "var(--color-accent)" : "var(--color-fg-faint)",
-            backgroundColor: filterTopic === null ? "var(--color-accent-soft)" : "transparent",
-          }}
-        >
-          all
-        </button>
-        {topicSet.slice(0, 12).map((k) => (
-          <button
-            key={`tk-${k}`}
-            type="button"
-            onClick={() => setFilterTopic(filterTopic === k ? null : k)}
-            className="rounded border px-1.5 py-0.5 text-[11px] font-mono"
-            style={{
-              borderColor: filterTopic === k ? TOPIC_COLORS[k % TOPIC_COLORS.length] : "var(--color-border)",
-              color: filterTopic === k ? TOPIC_COLORS[k % TOPIC_COLORS.length] : "var(--color-fg-faint)",
-              backgroundColor: filterTopic === k ? "var(--color-accent-soft)" : "transparent",
-            }}
-          >
-            t{k}
-          </button>
-        ))}
-        <span className="ml-3">label:</span>
-        <button
-          type="button"
-          onClick={() => setFilterLabel(null)}
-          className="rounded border px-1.5 py-0.5 text-[11px] font-mono"
-          style={{
-            borderColor: filterLabel === null ? "var(--color-accent)" : "var(--color-border)",
-            color: filterLabel === null ? "var(--color-accent)" : "var(--color-fg-faint)",
-            backgroundColor: filterLabel === null ? "var(--color-accent-soft)" : "transparent",
-          }}
-        >
-          all
-        </button>
-        {labelSet.slice(0, 16).map(([lid, lname]) => (
-          <button
-            key={`lab-${lid}`}
-            type="button"
-            onClick={() => setFilterLabel(filterLabel === lid ? null : lid)}
-            className="rounded border px-1.5 py-0.5 text-[10.5px] font-mono"
-            style={{
-              borderColor: filterLabel === lid ? "var(--color-accent)" : "var(--color-border)",
-              color: filterLabel === lid ? "var(--color-accent)" : "var(--color-fg-faint)",
-              backgroundColor: filterLabel === lid ? "var(--color-accent-soft)" : "transparent",
-            }}
-            title={lname}
-          >
-            {lid}
-          </button>
-        ))}
-      </div>
-
-      <div className="overflow-x-auto" style={{ maxHeight: 280 }}>
-        <table className="w-full text-[12px]" style={{ color: "var(--color-fg)" }}>
-          <thead className="sticky top-0" style={{ backgroundColor: "var(--color-panel)" }}>
-            <tr style={{ color: "var(--color-fg-faint)" }}>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">doc_id</th>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">label</th>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">dominant</th>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">θ stack (K = {K})</th>
-              <th className="text-right font-mono text-[11px] pb-1">θ_dominant</th>
-            </tr>
-          </thead>
-          <tbody>
-            {filtered.map(({ d, i }) => {
-              const isPicked = pickedIdx === i;
-              let acc = 0;
-              return (
-                <tr
-                  key={d.doc_id}
-                  onClick={() => onPick(i)}
-                  style={{
-                    borderTop: "1px solid var(--color-border)",
-                    backgroundColor: isPicked ? "var(--color-accent-soft)" : "transparent",
-                    cursor: "pointer",
-                  }}
-                >
-                  <td className="py-1 pr-3 font-mono">{d.doc_id}</td>
-                  <td className="py-1 pr-3 font-mono truncate" style={{ maxWidth: 100 }} title={d.label_name}>
-                    {d.label_name}
-                  </td>
-                  <td className="py-1 pr-3 font-mono">
-                    <span className="inline-block w-2 h-2 rounded-full mr-1 align-middle" style={{ backgroundColor: TOPIC_COLORS[d.topic_k_dominant % TOPIC_COLORS.length] }} />
-                    t{d.topic_k_dominant}
-                  </td>
-                  <td className="py-1 pr-3 w-[260px]">
-                    <div className="h-3 rounded overflow-hidden flex" style={{ backgroundColor: "var(--color-border)" }}>
-                      {d.theta_full?.map((p, kk) => {
-                        const w = p * 100;
-                        acc += w;
-                        return (
-                          <span
-                            key={`${d.doc_id}-${kk}`}
-                            style={{ width: `${w}%`, backgroundColor: TOPIC_COLORS[kk % TOPIC_COLORS.length] }}
-                            title={`topic ${kk}: ${(p * 100).toFixed(1)}%`}
-                          />
-                        );
-                      })}
-                      {acc < 99.9 ? <span style={{ width: `${100 - acc}%` }} /> : null}
-                    </div>
-                  </td>
-                  <td className="py-1 text-right font-mono">{(d.theta_k_at_dominant * 100).toFixed(1)}%</td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
-      <p className="text-[11px]" style={{ color: "var(--color-fg-faint)" }}>
-        {filtered.length} of {docCards.length} docs shown · click any row to drill in.
-      </p>
-    </div>
-  );
-}
-
-function DocDetailPanel({
-  doc,
-  marginalTheta,
-  topicViews,
-  topicToData,
-  selectedTopic,
-  setSelectedTopic,
-}: {
-  doc: import("@/api/client").DocumentCard;
-  marginalTheta: number[];
-  topicViews: import("@/api/client").TopicViews;
-  topicToData: import("@/api/client").TopicToData | null;
-  selectedTopic: number | null;
-  setSelectedTopic: (k: number | null) => void;
-}) {
-  const ranked = useMemo(() => {
-    return (doc.theta_full ?? [])
-      .map((p, k) => ({ k, p, marginal: marginalTheta[k] ?? 0 }))
-      .sort((a, b) => b.p - a.p);
-  }, [doc.theta_full, marginalTheta]);
-
-  const topicWords = topicViews.top_words_per_topic["lambda_0.5"] ?? [];
-
-  return (
-    <div className="space-y-5">
-      <div
-        className="rounded-lg border p-5"
-        style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)", boxShadow: "var(--color-shadow)" }}
-      >
-        <div className="grid sm:grid-cols-2 md:grid-cols-4 gap-4 mb-3">
-          <UnmixingStat label="doc_id" value={doc.doc_id} />
-          <UnmixingStat label="label" value={doc.label_name} />
-          <UnmixingStat label="dominant topic" value={`t${doc.topic_k_dominant}`} />
-          <UnmixingStat label="θ at dominant" value={`${(doc.theta_k_at_dominant * 100).toFixed(1)}%`} />
-        </div>
-
-        <h5 className="text-[11px] uppercase tracking-widest font-semibold mb-2" style={{ color: "var(--color-fg-faint)" }}>
-          θ ranked: this doc vs scene marginal · ratio (doc / marginal) highlights what the doc over-/under-weights
-        </h5>
-        <table className="w-full text-[12px]" style={{ color: "var(--color-fg)" }}>
-          <thead>
-            <tr style={{ color: "var(--color-fg-faint)" }}>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">topic</th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">doc θ</th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">marginal θ̄</th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">ratio</th>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">doc bar</th>
-              <th className="text-left font-mono text-[11px] pb-1">top words (λ=0.5)</th>
-            </tr>
-          </thead>
-          <tbody>
-            {ranked.map((row) => {
-              const colour = TOPIC_COLORS[row.k % TOPIC_COLORS.length];
-              const ratio = row.marginal > 0 ? row.p / row.marginal : 0;
-              const isSel = selectedTopic === row.k;
-              const words = topicWords?.[row.k] ?? [];
-              return (
-                <tr
-                  key={`r-${row.k}`}
-                  onClick={() => setSelectedTopic(isSel ? null : row.k)}
-                  style={{
-                    borderTop: "1px solid var(--color-border)",
-                    backgroundColor: isSel ? "var(--color-accent-soft)" : "transparent",
-                    cursor: "pointer",
-                  }}
-                >
-                  <td className="py-1 pr-3 font-mono">
-                    <span className="inline-block w-2 h-2 rounded-full mr-1 align-middle" style={{ backgroundColor: colour }} />
-                    t{row.k}
-                  </td>
-                  <td className="py-1 pr-3 text-right font-mono">{(row.p * 100).toFixed(1)}%</td>
-                  <td className="py-1 pr-3 text-right font-mono">{(row.marginal * 100).toFixed(1)}%</td>
-                  <td className="py-1 pr-3 text-right font-mono" style={{ color: ratio >= 2 ? "rgba(40,160,80,1)" : ratio <= 0.5 ? "rgba(214,39,40,1)" : "var(--color-fg-faint)" }}>
-                    {row.marginal > 0 ? ratio.toFixed(2) : "—"}×
-                  </td>
-                  <td className="py-1 pr-3 w-[140px]">
-                    <div className="h-2 rounded" style={{ backgroundColor: "var(--color-border)" }}>
-                      <div className="h-2 rounded" style={{ width: `${row.p * 100}%`, backgroundColor: colour }} />
-                    </div>
-                  </td>
-                  <td className="py-1 text-[11px] font-mono truncate" style={{ maxWidth: 220, color: "var(--color-fg-subtle)" }} title={words.slice(0, 8).map((w) => w.token).join(", ")}>
-                    {words.slice(0, 4).map((w) => w.token).join(", ")}
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
-
-      {topicToData ? (
-        <div
-          className="rounded-lg border p-4"
-          style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)", boxShadow: "var(--color-shadow)" }}
-        >
-          <h4 className="text-base font-semibold mb-1" style={{ color: "var(--color-fg)" }}>
-            P(label | dominant topic) — does the dominant topic predict the doc's label?
-          </h4>
-          <p className="text-[12px] mb-3" style={{ color: "var(--color-fg-faint)" }}>
-            For docs assigned to <code>t{doc.topic_k_dominant}</code>, here is the empirical label
-            distribution. If <strong>{doc.label_name}</strong> is the largest bar, the topic
-            "correctly routes" this doc's class.
-          </p>
-          <PerTopicLabelBars
-            row={topicToData.p_label_given_topic_dominant[doc.topic_k_dominant] ?? []}
-            highlightLabelId={doc.label_id}
-          />
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function PerTopicLabelBars({
-  row,
-  highlightLabelId,
-}: {
-  row: import("@/api/client").LabelCell[];
-  highlightLabelId: number;
-}) {
-  const sorted = [...row].sort((a, b) => b.p - a.p).slice(0, 12);
-  const max = sorted[0]?.p ?? 1;
-  return (
-    <div className="space-y-1">
-      {sorted.map((cell) => {
-        const isHi = cell.label_id === highlightLabelId;
-        const width = max > 0 ? (cell.p / max) * 100 : 0;
-        return (
-          <div key={`pl-${cell.label_id}`} className="flex items-baseline gap-2 text-[12px]">
-            <span className="font-mono w-32 truncate" style={{ color: isHi ? "var(--color-accent)" : "var(--color-fg)", fontWeight: isHi ? 600 : 400 }} title={cell.name}>
-              {isHi ? "★ " : ""}{cell.name}
-            </span>
-            <span className="font-mono text-[11px] w-12 text-right" style={{ color: "var(--color-fg-faint)" }}>
-              {(cell.p * 100).toFixed(1)}%
-            </span>
-            <div className="flex-1 h-2 rounded" style={{ backgroundColor: "var(--color-border)" }}>
-              <div className="h-2 rounded" style={{ width: `${width}%`, backgroundColor: isHi ? "var(--color-accent)" : "var(--color-fg-subtle)" }} />
-            </div>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-const RECIPE_IDS_V112 = ["V1","V2","V3","V4","V5","V6","V7","V8","V9","V10","V11","V12"];
-const SCHEMES = ["uniform", "quantile", "lloyd_max"] as const;
-const Q_VALUES = [8, 16, 32] as const;
-
-function RecipesTab({
-  sceneId,
-  isLoading,
-  error,
-  index,
-}: {
-  sceneId: string;
-  isLoading: boolean;
-  error: Error | null;
-  index: import("@/api/client").WordificationsIndex | null;
-}) {
-  const [recipe, setRecipe] = useState<string>("V1");
-  const [scheme, setScheme] = useState<typeof SCHEMES[number]>("uniform");
-  const [q, setQ] = useState<typeof Q_VALUES[number]>(8);
-
-  const detail = useQuery({
-    queryKey: ["wordification", sceneId, recipe, scheme, q],
-    queryFn: () => api.wordification(sceneId, recipe, scheme, q),
-    staleTime: 5 * 60_000,
-  });
-
-  if (isLoading) return <p style={{ color: "var(--color-fg-faint)" }}>Loading wordifications index…</p>;
-  if (error) {
-    return (
-      <div className="rounded-lg border p-6" style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)" }}>
-        <p style={{ color: "var(--color-warn)" }}>Could not load wordifications index.</p>
-        <p className="mt-2 text-sm" style={{ color: "var(--color-fg-faint)" }}>{error.message}</p>
-      </div>
-    );
-  }
-
-  // Build availability matrix from index
-  const available = new Set<string>();
-  if (index) {
-    const prefix = `${sceneId}_`;
-    for (const item of index.items) {
-      if (item.id.startsWith(prefix)) available.add(item.id.slice(prefix.length));
-    }
-  }
-
-  const sceneAvailable = available.size;
-
-  return (
-    <div className="space-y-6">
-      <div
-        className="rounded-xl border p-5 relative overflow-hidden"
-        style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)", boxShadow: "var(--color-shadow)" }}
-      >
-        <div
-          aria-hidden
-          className="absolute top-0 left-0 right-0 h-1"
-          style={{ background: "linear-gradient(90deg, rgba(40,160,80,1) 0%, rgba(56,189,248,1) 100%)" }}
-        />
-        <h3 className="text-lg font-semibold tracking-tight mt-1 mb-1" style={{ color: "var(--color-fg)" }}>
-          Wordification recipes · V1..V12 × scheme × Q
-        </h3>
-        <p className="text-[12.5px] mb-3" style={{ color: "var(--color-fg-faint)" }}>
-          {sceneId} has <strong>{sceneAvailable}</strong> precomputed wordification combos out of
-          the 12 × 3 × 3 = 108 grid. Topic-model artefacts (LDA φ, θ) are precomputed on the
-          canonical V1 / uniform / Q=8 recipe only. Other combos expose vocabulary statistics
-          (V_actual, doc-length distribution, top tokens) for inspection but not full LDA refits.
-        </p>
-
-        <div className="space-y-2">
-          <RecipeAxisPicker label="recipe" options={RECIPE_IDS_V112} value={recipe} onChange={(v) => setRecipe(v)} />
-          <RecipeAxisPicker label="scheme" options={SCHEMES as unknown as string[]} value={scheme} onChange={(v) => setScheme(v as typeof SCHEMES[number])} />
-          <RecipeAxisPicker label="Q" options={Q_VALUES.map(String)} value={String(q)} onChange={(v) => setQ(parseInt(v, 10) as typeof Q_VALUES[number])} />
-        </div>
-
-        <div className="mt-3 text-[11.5px] font-mono" style={{ color: available.has(`${recipe}_${scheme}_Q${q}`) ? "rgba(40,160,80,1)" : "var(--color-warn)" }}>
-          {available.has(`${recipe}_${scheme}_Q${q}`) ? "✓ this combo is available" : "✗ this combo not precomputed for this scene"}
-        </div>
-      </div>
-
-      {detail.isLoading ? (
-        <p style={{ color: "var(--color-fg-faint)" }}>Loading {recipe}/{scheme}/Q{q}…</p>
-      ) : detail.error ? (
-        <div className="rounded-lg border p-6" style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)" }}>
-          <p style={{ color: "var(--color-warn)" }}>Could not load this combo.</p>
-          <p className="mt-2 text-sm" style={{ color: "var(--color-fg-faint)" }}>{(detail.error as Error).message}</p>
-        </div>
-      ) : detail.data ? (
-        <RecipeDetailCard payload={detail.data} />
-      ) : null}
-    </div>
-  );
-}
-
-function RecipeAxisPicker({
-  label,
-  options,
-  value,
-  onChange,
-}: {
-  label: string;
-  options: string[];
-  value: string;
-  onChange: (v: string) => void;
-}) {
-  return (
-    <div className="flex items-baseline gap-2 flex-wrap">
-      <span className="text-[10.5px] uppercase tracking-widest font-semibold w-12" style={{ color: "var(--color-fg-faint)" }}>
-        {label}
-      </span>
-      <div className="flex flex-wrap gap-1">
-        {options.map((o) => {
-          const isActive = o === value;
-          return (
-            <button
-              key={o}
-              type="button"
-              onClick={() => onChange(o)}
-              className="rounded border px-2 py-0.5 text-[11.5px] font-mono"
-              style={{
-                borderColor: isActive ? "var(--color-accent)" : "var(--color-border)",
-                color: isActive ? "var(--color-accent)" : "var(--color-fg-subtle)",
-                backgroundColor: isActive ? "var(--color-accent-soft)" : "transparent",
-                fontWeight: isActive ? 600 : 400,
-              }}
-            >
-              {o}
-            </button>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-function RecipeDetailCard({ payload }: { payload: import("@/api/client").WordificationPayload }) {
-  const dl = payload.doc_length_distribution;
-  const top = payload.top_tokens_by_count ?? [];
-  return (
-    <div
-      className="rounded-lg border p-4"
-      style={{ borderColor: "var(--color-border)", backgroundColor: "var(--color-panel)", boxShadow: "var(--color-shadow)" }}
-    >
-      <h4 className="text-base font-semibold mb-1" style={{ color: "var(--color-fg)" }}>
-        {payload.recipe} · {payload.scheme} · Q={payload.Q}
-      </h4>
-      <div className="grid sm:grid-cols-2 md:grid-cols-4 gap-4 mt-3 mb-4">
-        <UnmixingStat label="docs (D)" value={payload.D.toLocaleString()} />
-        <UnmixingStat label="bands (B)" value={String(payload.B)} />
-        <UnmixingStat label="vocab full / actual" value={`${payload.V_full} / ${payload.V_actual}`} />
-        <UnmixingStat label="corpus entropy" value={payload.corpus_marginal_entropy_bits != null ? `${payload.corpus_marginal_entropy_bits.toFixed(3)} bits` : "—"} />
-      </div>
-      <div className="grid sm:grid-cols-2 gap-5">
-        <div>
-          <h5 className="text-[11px] uppercase tracking-widest font-semibold mb-2" style={{ color: "var(--color-fg-faint)" }}>
-            Document length distribution
-          </h5>
-          <div className="space-y-1 text-[12.5px] font-mono" style={{ color: "var(--color-fg)" }}>
-            <div>min/max: {dl.min} / {dl.max}</div>
-            <div>p25 / p50 / p75: {dl.p25.toFixed(0)} / {dl.p50.toFixed(0)} / {dl.p75.toFixed(0)}</div>
-            <div>mean ± std: {dl.mean.toFixed(2)} ± {dl.std.toFixed(2)}</div>
-            <div style={{ color: "var(--color-fg-faint)" }}>zero-token doc rate: {((payload.zero_token_doc_rate ?? 0) * 100).toFixed(2)}%</div>
-          </div>
-        </div>
-        {top.length > 0 ? (
-          <div>
-            <h5 className="text-[11px] uppercase tracking-widest font-semibold mb-2" style={{ color: "var(--color-fg-faint)" }}>
-              Top tokens by count
-            </h5>
-            <table className="w-full text-[12px]" style={{ color: "var(--color-fg)" }}>
-              <thead>
-                <tr style={{ color: "var(--color-fg-faint)" }}>
-                  <th className="text-left font-mono text-[10.5px] pb-1 pr-3">token</th>
-                  <th className="text-right font-mono text-[10.5px] pb-1 pr-3">count</th>
-                  <th className="text-right font-mono text-[10.5px] pb-1">p_global</th>
-                </tr>
-              </thead>
-              <tbody>
-                {top.slice(0, 10).map((t) => (
-                  <tr key={t.token} style={{ borderTop: "1px solid var(--color-border)" }}>
-                    <td className="py-1 pr-3 font-mono">{t.token}</td>
-                    <td className="py-1 pr-3 text-right font-mono">{t.count.toLocaleString()}</td>
-                    <td className="py-1 text-right font-mono text-[11px]" style={{ color: "var(--color-fg-faint)" }}>{(t.p_global * 100).toFixed(3)}%</td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-function QKExploreTab({
-  isLoading,
-  error,
-  sweep,
-}: {
-  isLoading: boolean;
-  error: Error | null;
-  sweep: import("@/api/client").LdaSweep | null;
-}) {
-  if (isLoading)
-    return <p style={{ color: "var(--color-fg-faint)" }}>Loading K-sweep…</p>;
-  if (error)
-    return (
-      <div
-        className="rounded-lg border p-6"
-        style={{
-          borderColor: "var(--color-border)",
-          backgroundColor: "var(--color-panel)",
-          boxShadow: "var(--color-shadow)",
-        }}
-      >
-        <p style={{ color: "var(--color-warn)" }}>Could not load lda_sweep.</p>
-        <p
-          className="mt-2 text-sm"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          {error.message}
-        </p>
-      </div>
-    );
-  if (!sweep || sweep.grid.length === 0) return null;
-
-  const Ks = sweep.grid.map((g) => g.K);
-  const perpMean = sweep.grid.map((g) => g.perplexity_test_mean);
-  const perpStd = sweep.grid.map((g) => g.perplexity_test_std);
-  const div = sweep.grid.map((g) => g.topic_diversity_mean);
-  const cos = sweep.grid.map((g) => g.matched_cosine_mean ?? 0);
-  const canonicalK = 12;
-  const recommendedK = sweep.recommended_K ?? null;
-
-  return (
-    <div className="space-y-5">
-      <div
-        className="rounded-lg border p-5"
-        style={{
-          borderColor: "var(--color-border)",
-          backgroundColor: "var(--color-panel)",
-          boxShadow: "var(--color-shadow)",
-        }}
-      >
-        <h4
-          className="text-base font-semibold mb-2"
-          style={{ color: "var(--color-fg)" }}
-        >
-          K-sweep · LDA model selection
-        </h4>
-        <p
-          className="text-sm mb-3"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          Precomputed sweep of K ∈ {`{${sweep.K_grid.join(", ")}}`} on{" "}
-          <span className="font-mono">{sweep.wordification}</span>, Q=
-          {sweep.quantization_scale}, averaged over {sweep.seeds.length} seeds
-          ({Math.round(sweep.train_fraction * 100)}% train split,{" "}
-          {sweep.samples_per_class.toLocaleString()} samples per class). The
-          Workspace currently fits at canonical K={canonicalK}; this tab shows
-          what the model would look like at other K values — live refit is{" "}
-          <em>not</em> supported because each combo would require ~minutes of
-          server compute.
-        </p>
-        {recommendedK !== null && (
-          <div
-            className="inline-flex items-center gap-2 rounded-md border px-3 py-1.5 mb-3 text-[12.5px]"
-            style={{
-              borderColor: "var(--color-accent)",
-              backgroundColor: "var(--color-accent-soft)",
-              color: "var(--color-accent)",
-            }}
-          >
-            <span className="font-semibold">Builder recommendation:</span>
-            <span className="font-mono">K={recommendedK}</span>
-            {sweep.recommendation_method && (
-              <span
-                className="font-mono text-[11px]"
-                style={{ color: "var(--color-fg-faint)" }}
-                title="Composite score: perplexity is rescaled to [0,1] across the K grid before being negated, then summed unweighted with npmi_mean and matched_cosine_mean (both already in their natural [0,1] range)."
-              >
-                ({sweep.recommendation_method})
-              </span>
-            )}
-            {recommendedK !== canonicalK && (
-              <span
-                className="text-[11px] italic"
-                style={{ color: "var(--color-fg-faint)" }}
-              >
-                — differs from canonical K={canonicalK}
-              </span>
-            )}
-          </div>
-        )}
-        {recommendedK !== null && recommendedK !== canonicalK && (
-          <p
-            className="text-[11.5px] mb-3"
-            style={{ color: "var(--color-fg-faint)" }}
-          >
-            <em>Note:</em> the canonical Workspace fit retains K={canonicalK}{" "}
-            for cross-scene comparability. The composite score above is a
-            model-selection observation, not a mandate — the rest of the
-            app (topics, raster, applydoc, etc.) keeps reading the K=
-            {canonicalK} canonical fit.
-          </p>
-        )}
-        <div className="grid lg:grid-cols-3 gap-4">
-          <SweepCurveCard
-            title="Perplexity (test)"
-            subtitle="lower is better; ↓ shows the model captures the held-out word distribution"
-            xs={Ks}
-            ys={perpMean}
-            stds={perpStd}
-            highlightX={canonicalK}
-          />
-          <SweepCurveCard
-            title="Topic diversity"
-            subtitle="fraction of unique top-words across topics; ↓ as K grows = topics overlap more"
-            xs={Ks}
-            ys={div}
-            highlightX={canonicalK}
-          />
-          <SweepCurveCard
-            title="Matched cosine"
-            subtitle="seed-to-seed mean cosine of best-matched topics; ↑ = more stable fit"
-            xs={Ks}
-            ys={cos}
-            highlightX={canonicalK}
-          />
-        </div>
-      </div>
-
-      <div
-        className="rounded-lg border p-5"
-        style={{
-          borderColor: "var(--color-border)",
-          backgroundColor: "var(--color-panel)",
-          boxShadow: "var(--color-shadow)",
-        }}
-      >
-        <h4
-          className="text-base font-semibold mb-2"
-          style={{ color: "var(--color-fg)" }}
-        >
-          Per-K detail
-        </h4>
-        <table className="w-full text-[12.5px]" style={{ color: "var(--color-fg)" }}>
-          <thead>
-            <tr style={{ color: "var(--color-fg-faint)" }}>
-              <th className="text-left font-mono text-[11px] pb-1 pr-3">K</th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">
-                perplexity mean
-              </th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">
-                perplexity std
-              </th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">
-                topic diversity
-              </th>
-              <th className="text-right font-mono text-[11px] pb-1 pr-3">
-                matched cosine mean
-              </th>
-              <th className="text-right font-mono text-[11px] pb-1">
-                matched cosine min
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {sweep.grid.map((g) => {
-              const isCanon = g.K === canonicalK;
-              const isRec = recommendedK !== null && g.K === recommendedK;
-              return (
-                <tr
-                  key={g.K}
-                  style={{
-                    borderTop: "1px solid var(--color-border)",
-                    backgroundColor:
-                      isCanon || isRec
-                        ? "var(--color-accent-soft)"
-                        : "transparent",
-                  }}
-                >
-                  <td className="py-1 pr-3 font-mono">
-                    K={g.K}
-                    {isCanon ? " ★" : ""}
-                    {isRec && !isCanon ? " ●" : ""}
-                  </td>
-                  <td className="py-1 pr-3 text-right font-mono">
-                    {g.perplexity_test_mean.toFixed(3)}
-                  </td>
-                  <td
-                    className="py-1 pr-3 text-right font-mono"
-                    style={{ color: "var(--color-fg-faint)" }}
-                  >
-                    ±{g.perplexity_test_std.toFixed(3)}
-                  </td>
-                  <td className="py-1 pr-3 text-right font-mono">
-                    {g.topic_diversity_mean.toFixed(3)}
-                  </td>
-                  <td className="py-1 pr-3 text-right font-mono">
-                    {g.matched_cosine_mean != null
-                      ? g.matched_cosine_mean.toFixed(3)
-                      : "—"}
-                  </td>
-                  <td className="py-1 text-right font-mono">
-                    {g.matched_cosine_min != null
-                      ? g.matched_cosine_min.toFixed(3)
-                      : "—"}
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-        <p
-          className="text-[11.5px] mt-3"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          ★ = canonical K used everywhere else in the Workspace.{" "}
-          {recommendedK !== null && recommendedK !== canonicalK
-            ? "● = builder-recommended K from sweep aggregation. "
-            : ""}
-          For Q
-          sensitivity (vocabulary granularity), see the{" "}
-          <span className="font-mono">robust</span> tab.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-function SweepCurveCard({
-  title,
-  subtitle,
-  xs,
-  ys,
-  stds,
-  highlightX,
-}: {
-  title: string;
-  subtitle: string;
-  xs: number[];
-  ys: number[];
-  stds?: number[];
-  highlightX?: number;
-}) {
-  const W = 240;
-  const H = 130;
-  const PAD = 28;
-  const xMin = Math.min(...xs);
-  const xMax = Math.max(...xs);
-  const yLo = Math.min(...ys.map((y, i) => y - (stds?.[i] ?? 0)));
-  const yHi = Math.max(...ys.map((y, i) => y + (stds?.[i] ?? 0)));
-  const pad = (yHi - yLo) * 0.1 || 0.05;
-  const yMin = yLo - pad;
-  const yMax = yHi + pad;
-  const px = (x: number) =>
-    PAD + ((x - xMin) / (xMax - xMin)) * (W - 2 * PAD);
-  const py = (y: number) =>
-    H - PAD - ((y - yMin) / (yMax - yMin)) * (H - 2 * PAD);
-  const linePts = xs.map((x, i) => `${px(x).toFixed(1)},${py(ys[i]!).toFixed(1)}`).join(" ");
-  const bandPts = stds
-    ? xs
-        .map(
-          (x, i) => `${px(x).toFixed(1)},${py(ys[i]! - stds[i]!).toFixed(1)}`,
-        )
-        .concat(
-          [...xs]
-            .reverse()
-            .map(
-              (x, i) =>
-                `${px(x).toFixed(1)},${py(ys[xs.length - 1 - i]! + stds[xs.length - 1 - i]!).toFixed(1)}`,
-            ),
-        )
-        .join(" ")
-    : null;
-  return (
-    <div
-      className="rounded-md border p-3"
-      style={{
-        borderColor: "var(--color-border)",
-        backgroundColor: "var(--color-bg)",
-      }}
-    >
-      <div
-        className="text-[12.5px] font-semibold mb-0.5"
-        style={{ color: "var(--color-fg)" }}
-      >
-        {title}
-      </div>
-      <div
-        className="text-[11px] mb-2 leading-snug"
-        style={{ color: "var(--color-fg-faint)" }}
-      >
-        {subtitle}
-      </div>
-      <svg
-        width="100%"
-        viewBox={`0 0 ${W} ${H}`}
-        xmlns="http://www.w3.org/2000/svg"
-        role="img"
-        aria-label={title}
-        style={{ color: "var(--color-fg)", display: "block" }}
-      >
-        <line
-          x1={PAD}
-          y1={H - PAD}
-          x2={W - PAD}
-          y2={H - PAD}
-          stroke="var(--color-border)"
-        />
-        <line
-          x1={PAD}
-          y1={PAD}
-          x2={PAD}
-          y2={H - PAD}
-          stroke="var(--color-border)"
-        />
-        {bandPts && (
-          <polygon
-            points={bandPts}
-            fill="var(--color-accent)"
-            opacity="0.18"
-          />
-        )}
-        <polyline
-          points={linePts}
-          fill="none"
-          stroke="var(--color-accent)"
-          strokeWidth="1.5"
-        />
-        {xs.map((x, i) => (
-          <circle
-            key={i}
-            cx={px(x)}
-            cy={py(ys[i]!)}
-            r={x === highlightX ? 4 : 2.5}
-            fill="var(--color-accent)"
-            stroke={
-              x === highlightX ? "var(--color-fg)" : "var(--color-accent)"
-            }
-            strokeWidth={x === highlightX ? 1.5 : 0}
-          />
-        ))}
-        {xs.map((x, i) => (
-          <text
-            key={`xl-${i}`}
-            x={px(x)}
-            y={H - PAD + 12}
-            textAnchor="middle"
-            fontSize="10"
-            fontFamily="ui-monospace, monospace"
-            fill="var(--color-fg-faint)"
-          >
-            {x}
-          </text>
-        ))}
-        <text
-          x={PAD - 4}
-          y={py(yMax)}
-          textAnchor="end"
-          fontSize="10"
-          fontFamily="ui-monospace, monospace"
-          fill="var(--color-fg-faint)"
-        >
-          {yMax.toFixed(2)}
-        </text>
-        <text
-          x={PAD - 4}
-          y={py(yMin)}
-          textAnchor="end"
-          fontSize="10"
-          fontFamily="ui-monospace, monospace"
-          fill="var(--color-fg-faint)"
-        >
-          {yMin.toFixed(2)}
-        </text>
-        <text
-          x={W / 2}
-          y={H - 4}
-          textAnchor="middle"
-          fontSize="10"
-          fill="var(--color-fg-faint)"
-        >
-          K
-        </text>
-      </svg>
-    </div>
-  );
-}
-
-function BandMaskTab({
-  sceneId,
-  isLoading,
-  error,
-  index,
-}: {
-  sceneId: string;
-  isLoading: boolean;
-  error: Error | null;
-  index: import("@/api/client").BandMaskIndex | null;
-}) {
-  const [maskId, setMaskId] = useState<string | null>(null);
-  const summaryQ = useQuery({
-    queryKey: ["band-mask-summary", sceneId, maskId],
-    queryFn: () => api.bandMaskSummary(sceneId, maskId!),
-    enabled: maskId !== null,
-    staleTime: 30 * 60_000,
-  });
-  // Gate the canonical-comparison fetch on the user having selected a
-  // band-mask card. Most users browsing the bandmask tab will be reading
-  // the BandMaskComparisonOverview table that is rendered immediately,
-  // which IS the comparison payload — so we keep the fetch eager here.
-  // sceneId stays in the query key so a SceneQuickSwitch invalidates
-  // the cache rather than silently re-using a stale dataset's table.
-  const comparisonQ = useQuery({
-    queryKey: ["band-masks-comparison", sceneId],
-    queryFn: () => api.bandMasksCanonicalComparison(),
-    enabled: !!sceneId,
-    staleTime: 30 * 60_000,
-  });
-
-  if (isLoading)
-    return <p style={{ color: "var(--color-fg-faint)" }}>Loading band-mask index…</p>;
-  if (error)
-    return (
-      <div
-        className="rounded-lg border p-6"
-        style={{
-          borderColor: "var(--color-border)",
-          backgroundColor: "var(--color-panel)",
-          boxShadow: "var(--color-shadow)",
-        }}
-      >
-        <p style={{ color: "var(--color-warn)" }}>
-          Could not load /api/band-masks: {error.message}
-        </p>
-        <p
-          className="mt-2 text-sm"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          Run <span className="font-mono">scripts/local.* build-band-masked-topic-models</span> to generate the index locally.
-        </p>
-      </div>
-    );
-  if (!index) return null;
-
-  const sceneEntries = index.entries.filter((e) => e.scene_id === sceneId);
-  const maskDefs = index.mask_definitions;
-
-  return (
-    <div className="space-y-5">
-      <div
-        className="rounded-lg border p-5"
-        style={{
-          borderColor: "var(--color-border)",
-          backgroundColor: "var(--color-panel)",
-          boxShadow: "var(--color-shadow)",
-        }}
-      >
-        <h4
-          className="text-base font-semibold mb-2"
-          style={{ color: "var(--color-fg)" }}
-        >
-          Step 8 · band-mask sweep
-        </h4>
-        <p
-          className="text-sm mb-3"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          Four band-restricted LDA fits per labelled scene, each one a fresh
-          canonical-K refit on a band-restricted corpus. The user can ask:
-          "what would the topics look like if we only saw the VNIR / SWIR /
-          non-water-corrupted / top-50-discriminative bands?". Hyperparameters
-          mirror the canonical fit (V1 band-frequency, online VB, α=0.45,
-          η=0.20, seed=42) so only the band-selection axis differs. The
-          summary below ships precomputed; click any mask to drill into the
-          per-topic detail.
-        </p>
-        <ul className="grid sm:grid-cols-2 lg:grid-cols-4 gap-3">
-          {Object.entries(maskDefs).map(([id, def]) => {
-            const entry = sceneEntries.find((e) => e.mask_id === id);
-            const skipped = entry?.skipped ?? false;
-            const isSel = maskId === id;
-            return (
-              <li key={id}>
-                <button
-                  type="button"
-                  onClick={() => !skipped && setMaskId(isSel ? null : id)}
-                  disabled={skipped}
-                  className="w-full text-left rounded-md border p-3 transition-all"
-                  style={{
-                    borderColor: isSel
-                      ? "var(--color-accent)"
-                      : "var(--color-border)",
-                    backgroundColor: isSel
-                      ? "var(--color-accent-soft)"
-                      : "var(--color-bg)",
-                    opacity: skipped ? 0.4 : 1,
-                    cursor: skipped ? "not-allowed" : "pointer",
-                  }}
-                >
-                  <div
-                    className="font-mono text-[13px] font-semibold mb-0.5"
-                    style={{
-                      color: isSel ? "var(--color-accent)" : "var(--color-fg)",
-                    }}
-                  >
-                    {def.label}
-                  </div>
-                  <div
-                    className="text-[11.5px] leading-snug mb-1"
-                    style={{ color: "var(--color-fg-faint)" }}
-                  >
-                    {def.description}
-                  </div>
-                  {skipped ? (
-                    <div
-                      className="text-[11px] font-mono"
-                      style={{ color: "var(--color-warn)" }}
-                    >
-                      skipped: {entry?.reason}
-                    </div>
-                  ) : entry ? (
-                    <div
-                      className="grid grid-cols-2 gap-x-3 gap-y-0.5 text-[11.5px] font-mono mt-1.5"
-                      style={{ color: "var(--color-fg-subtle)" }}
-                    >
-                      <span>
-                        <span style={{ color: "var(--color-fg-faint)" }}>
-                          bands:{" "}
-                        </span>
-                        <span style={{ color: "var(--color-fg)" }}>
-                          {entry.n_bands_kept}/{entry.n_bands_full}
-                        </span>
-                      </span>
-                      <span>
-                        <span style={{ color: "var(--color-fg-faint)" }}>
-                          K:{" "}
-                        </span>
-                        <span style={{ color: "var(--color-fg)" }}>
-                          {entry.topic_count}
-                        </span>
-                      </span>
-                      <span>
-                        <span style={{ color: "var(--color-fg-faint)" }}>
-                          ppl:{" "}
-                        </span>
-                        <span style={{ color: "var(--color-fg)" }}>
-                          {entry.perplexity_train?.toFixed(2) ?? "—"}
-                        </span>
-                      </span>
-                      <span>
-                        <span style={{ color: "var(--color-fg-faint)" }}>
-                          ARI:{" "}
-                        </span>
-                        <span style={{ color: "var(--color-fg)" }}>
-                          {entry.ari_dominant_vs_label?.toFixed(3) ?? "—"}
-                        </span>
-                      </span>
-                    </div>
-                  ) : (
-                    <div
-                      className="text-[11px]"
-                      style={{ color: "var(--color-fg-faint)" }}
-                    >
-                      not built for this scene
-                    </div>
-                  )}
-                </button>
-              </li>
-            );
-          })}
-        </ul>
-        <p
-          className="text-[11px] mt-3"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          ARI = adjusted-Rand-index of <span className="font-mono">argmax θ_d</span> vs ground-truth label. Lower perplexity is better; higher ARI is better. Compare to the canonical (no-mask) fit on the <span className="font-mono">topics</span> / <span className="font-mono">topiclabel</span> tabs.
-        </p>
-      </div>
-
-      {comparisonQ.data && (
-        <BandMaskComparisonOverview
-          sceneId={sceneId}
-          comparison={comparisonQ.data}
-        />
-      )}
-
-      {maskId && summaryQ.data && (
-        <BandMaskDetailCard
-          summary={summaryQ.data}
-          comparison={
-            comparisonQ.data?.entries.find(
-              (e) => e.scene_id === sceneId && e.mask_id === maskId,
-            ) ?? null
-          }
-        />
-      )}
-      {maskId && summaryQ.isLoading && (
-        <p style={{ color: "var(--color-fg-faint)" }}>Loading {maskId} summary…</p>
-      )}
-      {maskId && summaryQ.error && (
-        <p style={{ color: "var(--color-warn)" }}>
-          Could not load {maskId} summary: {(summaryQ.error as Error).message}
-        </p>
-      )}
-    </div>
-  );
-}
-
-function BandMaskDetailCard({
-  summary,
-  comparison,
-}: {
-  summary: import("@/api/client").BandMaskSummary;
-  comparison: import("@/api/client").BandMaskComparisonEntry | null;
-}) {
-  const K = summary.topic_count;
-  return (
-    <div
-      className="rounded-lg border p-5"
-      style={{
-        borderColor: "var(--color-border)",
-        backgroundColor: "var(--color-panel)",
-        boxShadow: "var(--color-shadow)",
-      }}
-    >
-      <header className="mb-3">
-        <h4
-          className="text-base font-semibold"
-          style={{ color: "var(--color-fg)" }}
-        >
-          {summary.mask_label}
-        </h4>
-        <p
-          className="text-sm mt-1"
-          style={{ color: "var(--color-fg-faint)" }}
-        >
-          {summary.mask_description}
-        </p>
-      </header>
-      <div className="grid sm:grid-cols-3 md:grid-cols-6 gap-3 mb-4">
-        <BandMaskStat label="bands kept" value={`${summary.n_bands_kept}/${summary.n_bands_full}`} />
-        <BandMaskStat label="K" value={String(K)} />
-        <BandMaskStat label="D" value={summary.document_count.toLocaleString()} />
-        <BandMaskStat label="V" value={summary.vocabulary_size.toLocaleString()} />
-        <BandMaskStat label="perplexity (train)" value={summary.perplexity_train.toFixed(3)} />
-        <BandMaskStat label="ARI vs label" value={summary.ari_dominant_vs_label.toFixed(4)} />
-      </div>
-      <div
-        className="rounded-md border p-3 text-[11.5px] mb-3"
-        style={{
-          borderColor: "var(--color-border)",
-          backgroundColor: "var(--color-bg)",
-          color: "var(--color-fg-faint)",
-        }}
-      >
-        Wavelength range kept: {summary.wavelengths_nm_kept_first_last[0].toFixed(1)} – {summary.wavelengths_nm_kept_first_last[1].toFixed(1)} nm
-        {" · "}
-        first ten bands kept: <span className="font-mono">{summary.kept_band_indices.slice(0, 10).join(", ")}</span>
-        {summary.kept_band_indices.length > 10 ? ", …" : ""}
-      </div>
-      <div className="grid lg:grid-cols-2 gap-5">
-        <div>
-          <h5
-            className="text-[12px] uppercase tracking-widest font-semibold mb-2"
-            style={{ color: "var(--color-fg-faint)" }}
-          >
-            Top words per topic (λ=0.5)
-          </h5>
-          <ul className="space-y-1.5 text-[12px]">
-            {summary.top_words_per_topic_lambda_05.map((words, k) => {
-              const color =
-                TOPIC_COLORS[k % TOPIC_COLORS.length] ?? "#0ea5e9";
-              return (
-                <li
-                  key={k}
-                  className="grid grid-cols-[60px_1fr] gap-2 items-baseline"
-                  style={{ color: "var(--color-fg-subtle)" }}
-                >
-                  <span
-                    className="inline-flex items-center gap-1.5 font-mono"
-                    style={{ color: "var(--color-fg)" }}
-                  >
-                    <span
-                      aria-hidden
-                      className="inline-block w-2 h-2 rounded-sm"
-                      style={{ backgroundColor: color }}
-                    />
-                    t{k + 1}
-                  </span>
-                  <span className="font-mono text-[11.5px]">
-                    {words.slice(0, 8).join(" · ")}
-                  </span>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
-        <div>
-          <h5
-            className="text-[12px] uppercase tracking-widest font-semibold mb-2"
-            style={{ color: "var(--color-fg-faint)" }}
-          >
-            P(label | topic dominant)
-          </h5>
-          <ul className="space-y-1.5 text-[12px]">
-            {summary.p_label_given_topic_dominant.map((labels, k) => {
-              const sorted = [...labels].sort((a, b) => b.p - a.p).slice(0, 3);
-              const color =
-                TOPIC_COLORS[k % TOPIC_COLORS.length] ?? "#0ea5e9";
-              return (
-                <li
-                  key={k}
-                  className="grid grid-cols-[60px_1fr] gap-2 items-baseline"
-                  style={{ color: "var(--color-fg-subtle)" }}
-                >
-                  <span
-                    className="inline-flex items-center gap-1.5 font-mono"
-                    style={{ color: "var(--color-fg)" }}
-                  >
-                    <span
-                      aria-hidden
-                      className="inline-block w-2 h-2 rounded-sm"
-                      style={{ backgroundColor: color }}
-                    />
-                    t{k + 1}
-                  </span>
-                  <span className="font-mono text-[11.5px]">
-                    {sorted
-                      .map((l) => `${l.name} ${(l.p * 100).toFixed(0)}%`)
-                      .join(" · ")}
-                  </span>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
-      </div>
-
-      {comparison && !comparison.skipped && (
-        <div
-          className="mt-5 rounded-md border p-3"
-          style={{
-            borderColor: "var(--color-border)",
-            backgroundColor: "var(--color-bg)",
-          }}
-        >
-          <div
-            className="text-[11px] uppercase tracking-wider mb-2"
-            style={{ color: "var(--color-fg-faint)" }}
-          >
-            vs canonical fit (cycle 127 post-processor)
-          </div>
-          <div className="grid sm:grid-cols-2 lg:grid-cols-4 gap-3 mb-3">
-            <BandMaskStat
-              label="paired ARI dominant"
-              value={
-                comparison.paired_ari_dominant_topics != null
-                  ? comparison.paired_ari_dominant_topics.toFixed(4)
-                  : "—"
-              }
-            />
-            <BandMaskStat
-              label="swap rate (Hungarian)"
-              value={
-                comparison.swap_rate_under_hungarian_alignment != null
-                  ? `${(comparison.swap_rate_under_hungarian_alignment * 100).toFixed(1)}%`
-                  : "—"
-              }
-            />
-            <BandMaskStat
-              label="KL P(L|t) mean"
-              value={
-                comparison.kl_p_label_given_topic_mean != null
-                  ? comparison.kl_p_label_given_topic_mean.toFixed(2)
-                  : "—"
-              }
-            />
-            <BandMaskStat
-              label="KL P(L|t) max"
-              value={
-                comparison.kl_p_label_given_topic_max != null
-                  ? comparison.kl_p_label_given_topic_max.toFixed(2)
-                  : "—"
-              }
-            />
-          </div>
-          <p
-            className="text-[11.5px] mb-2"
-            style={{ color: "var(--color-fg-faint)" }}
-          >
-            Paired ARI measures how much the dominant-topic assignments shift
-            under the mask (1.0 = identical; 0.0 = unrelated). Swap rate counts,
-            under a Hungarian alignment of topic ids, the fraction of pixels
-            whose dominant topic changed. KL is the mean / max KL divergence
-            between canonical P(L|t) and masked P(L|t) for the Hungarian-aligned
-            topic pairs (higher = the masked fit assigns labels very differently
-            to its matched topics).
-          </p>
-          {comparison.hungarian_assignment && (
-            <div className="text-[11.5px]">
-              <div
-                className="text-[10.5px] uppercase tracking-wider mb-1"
-                style={{ color: "var(--color-fg-faint)" }}
-              >
-                Hungarian topic-id alignment (canonical → masked)
-              </div>
-              <div className="flex flex-wrap gap-2 font-mono">
-                {Object.entries(comparison.hungarian_assignment)
-                  .sort((a, b) => Number(a[0]) - Number(b[0]))
-                  .map(([can, mas]) => (
-                    <span
-                      key={can}
-                      className="rounded-sm border px-1.5 py-0.5"
-                      style={{
-                        borderColor: "var(--color-border)",
-                        color: "var(--color-fg)",
-                      }}
-                    >
-                      t{Number(can) + 1}→t{(mas as number) + 1}
-                    </span>
-                  ))}
-              </div>
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function BandMaskComparisonOverview({
-  sceneId,
-  comparison,
-}: {
-  sceneId: string;
-  comparison: import("@/api/client").BandMaskCanonicalComparison;
-}) {
-  const sceneEntries = comparison.entries.filter(
-    (e) => e.scene_id === sceneId && !e.skipped,
-  );
-  if (sceneEntries.length === 0) return null;
-  return (
-    <div
-      className="rounded-lg border p-5"
-      style={{
-        borderColor: "var(--color-border)",
-        backgroundColor: "var(--color-panel)",
-        boxShadow: "var(--color-shadow)",
-      }}
-    >
-      <h4
-        className="text-base font-semibold mb-2"
-        style={{ color: "var(--color-fg)" }}
-      >
-        Canonical-vs-masked comparison (all masks for this scene)
-      </h4>
-      <p
-        className="text-[12.5px] mb-3"
-        style={{ color: "var(--color-fg-faint)" }}
-      >
-        Cycle-127 read-only post-processor. Paired ARI = adjusted-Rand-index
-        of the dominant-topic map under the mask vs the canonical fit on
-        pixels labelled in both. Swap rate = fraction of pixels whose
-        Hungarian-aligned dominant topic differs. Lower paired ARI / higher
-        swap rate ⇒ the masked fit yields a substantially different topic
-        assignment than the canonical fit.
-      </p>
-      <table
-        className="w-full text-[12px]"
-        style={{ color: "var(--color-fg)" }}
-      >
-        <thead>
-          <tr style={{ color: "var(--color-fg-faint)" }}>
-            <th className="text-left font-mono text-[11px] pb-1 pr-3">mask</th>
-            <th className="text-right font-mono text-[11px] pb-1 pr-3">paired ARI</th>
-            <th className="text-right font-mono text-[11px] pb-1 pr-3">swap rate</th>
-            <th className="text-right font-mono text-[11px] pb-1 pr-3">KL mean</th>
-            <th className="text-right font-mono text-[11px] pb-1 pr-3">KL max</th>
-            <th className="text-right font-mono text-[11px] pb-1 pr-3">ARI vs label</th>
-            <th className="text-right font-mono text-[11px] pb-1">ppl</th>
-          </tr>
-        </thead>
-        <tbody>
-          {sceneEntries.map((e) => (
-            <tr
-              key={e.mask_id}
-              style={{ borderTop: "1px solid var(--color-border)" }}
-            >
-              <td className="py-1 pr-3 font-mono">{e.mask_id}</td>
-              <td className="py-1 pr-3 text-right font-mono">
-                {e.paired_ari_dominant_topics != null
-                  ? e.paired_ari_dominant_topics.toFixed(4)
-                  : "—"}
-              </td>
-              <td className="py-1 pr-3 text-right font-mono">
-                {e.swap_rate_under_hungarian_alignment != null
-                  ? `${(e.swap_rate_under_hungarian_alignment * 100).toFixed(1)}%`
-                  : "—"}
-              </td>
-              <td className="py-1 pr-3 text-right font-mono">
-                {e.kl_p_label_given_topic_mean != null
-                  ? e.kl_p_label_given_topic_mean.toFixed(2)
-                  : "—"}
-              </td>
-              <td className="py-1 pr-3 text-right font-mono">
-                {e.kl_p_label_given_topic_max != null
-                  ? e.kl_p_label_given_topic_max.toFixed(2)
-                  : "—"}
-              </td>
-              <td className="py-1 pr-3 text-right font-mono">
-                {e.ari_dominant_vs_label_masked != null
-                  ? e.ari_dominant_vs_label_masked.toFixed(3)
-                  : "—"}
-              </td>
-              <td className="py-1 text-right font-mono">
-                {e.perplexity_train_masked != null
-                  ? e.perplexity_train_masked.toFixed(2)
-                  : "—"}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-function BandMaskStat({ label, value }: { label: string; value: string }) {
-  return (
-    <div
-      className="rounded-md border p-2.5"
-      style={{
-        borderColor: "var(--color-border)",
-        backgroundColor: "var(--color-bg)",
-      }}
-    >
-      <div
-        className="text-[10.5px] uppercase tracking-wider"
-        style={{ color: "var(--color-fg-faint)" }}
-      >
-        {label}
-      </div>
-      <div
-        className="text-[14px] font-mono mt-0.5"
-        style={{ color: "var(--color-fg)" }}
-      >
-        {value}
-      </div>
-    </div>
-  );
-}
